# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Charity API",
        "description": "توثيق RESTful API لنظام الجمعية الخيرية",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "https://Al-Talahom-Charity.com/api",
            "description": "Production API Server"
        }
    ],
    "paths": {
        "/api/initiatives": {
            "get": {
                "tags": [
                    "Initiatives"
                ],
                "summary": "جلب كل المبادرات",
                "operationId": "396e83d95b2811aeebd0e1e45c66a38d",
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InitiativeResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Initiatives"
                ],
                "summary": "إنشاء مبادرة جديدة",
                "operationId": "33cbee379d054f097f21789473c5d638",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "مبادرة خيرية"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "تفاصيل المبادرة"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم الإنشاء",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InitiativeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/initiatives/{id}": {
            "get": {
                "tags": [
                    "Initiatives"
                ],
                "summary": "عرض مبادرة واحدة",
                "operationId": "a334eeb3c0c32bbb46b5cd3c0b22491c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InitiativeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            },
            "put": {
                "tags": [
                    "Initiatives"
                ],
                "summary": "تحديث مبادرة",
                "operationId": "5f91208ca1746ba658e164b5f583b633",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "مبادرة محدثة"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "تفاصيل محدثة"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم التحديث",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InitiativeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Initiatives"
                ],
                "summary": "حذف مبادرة",
                "operationId": "509336080f99c63119313e62665bd25e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم الحذف"
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            }
        },
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "جلب كل المشاريع",
                "operationId": "fbc21feeb3df8a2cb519efb6480ae7d9",
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "إنشاء مشروع جديد",
                "operationId": "a9d5bd5914cf75eaebafcf8b9e320dcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "مشروع خيري"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "تفاصيل المشروع"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم الإنشاء",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "عرض مشروع واحد",
                "operationId": "865085b82af27d647cf1f2a772aa7927",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نجاح العملية",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "تحديث مشروع",
                "operationId": "fe600b7ffce78456bce67c51b9eca51d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "مشروع محدث"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "تحديث التفاصيل"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "inactive"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم التحديث",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "حذف مشروع",
                "operationId": "d81ac2b0b51fa7143c69f1a5f4c27eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم الحذف"
                    },
                    "404": {
                        "description": "غير موجود"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Project": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "مشروع خيري"
                    },
                    "description": {
                        "type": "string",
                        "example": "تفاصيل المشروع"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "ProjectResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Projects retrieved successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    }
                },
                "type": "object"
            },
            "Initiative": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "مبادرة خيرية"
                    },
                    "description": {
                        "type": "string",
                        "example": "تفاصيل المبادرة"
                    }
                },
                "type": "object"
            },
            "InitiativeResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Initiatives retrieved successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Initiative"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}